move vtable code into folder
