+++ packages/lix-sdk/src/state/transaction/insert-transaction-state.ts
@@ -3,7 +3,7 @@ import { executeSync } from "../../database/execute-sync.js";
 import { uuidV7 } from "../../deterministic/index.js";
 import type { Lix } from "../../lix/open-lix.js";
 import type { LixInternalDatabaseSchema } from "../../database/schema.js";
-import type { NewStateAllRow, StateAllRow } from "../schema.js";
+import type { NewStateAllRow, StateAllRow } from "../index.js";
 
 type NewTransactionStateRow = Omit<NewStateAllRow, "snapshot_content"> & {
 	snapshot_content: string | null;
