+++ packages/lix-sdk/src/state/vtable/index.ts
@@ -1,0 +1,4 @@
+export { applyStateVTable } from "./vtable.js";
+export { commit } from "./commit.js";
+export { validateStateMutation } from "./validate-state-mutation.js";
+export { serializeStatePk, parseStatePk } from "./primary-key.js";
